[Oztechan/CCC#2625] Move Slides to new RootViewLogic
